+++ src/Umbraco.Abstractions/Search/IUmbracoTreeSearcherFields.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 
 namespace Umbraco.Web.Search
-{
+{ 
     /// <summary>
     /// Used to propagate hardcoded internal Field lists
     /// </summary>
