+++ src/Umbraco.Examine.Lucene/ExamineLuceneFinalComposer.cs
@@ -1,0 +1,13 @@
+﻿using Umbraco.Core;
+using Umbraco.Core.Composing;
+
+namespace Umbraco.Examine
+{
+    // examine's Lucene final composer composes after all user composers
+    // and *also* after ICoreComposer (in case IUserComposer is disabled)
+    [ComposeAfter(typeof(IUserComposer))]
+    [ComposeAfter(typeof(ICoreComposer))]
+    [RuntimeLevel(MinLevel = RuntimeLevel.Run)]
+    public class ExamineLuceneFinalComposer : ComponentComposer<ExamineLuceneFinalComponent>
+    { }
+}
