+++ src/Umbraco.Examine.Lucene/UmbracoIndexesCreator.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using Umbraco.Core.Logging;
 using Umbraco.Core.Services;
 using Umbraco.Examine;
@@ -9,7 +8,6 @@ using Examine;
 using Umbraco.Core;
 using Umbraco.Core.Composing;
 using Umbraco.Core.IO;
-using Umbraco.Web.PublishedCache.NuCache;
 
 namespace Umbraco.Web.Search
 {
