+++ src/Umbraco.Examine/IUmbracoIndexesCreator.cs
@@ -1,8 +1,4 @@
-﻿using System.Collections.Generic;
-using Examine;
-using Umbraco.Examine;
-
-namespace Umbraco.Web.Search
+﻿namespace Umbraco.Examine
 {
     /// <inheritdoc />
     /// <summary>
