+++ src/Umbraco.Web/Search/BackgroundIndexRebuilder.cs
@@ -118,7 +118,6 @@ namespace Umbraco.Web.Search
                 if (_waitMilliseconds > 0)
                     Thread.Sleep(_waitMilliseconds);
 
-                _indexRebuilder.ExamineManager.ConfigureIndexes(_mainDom, _logger);
                 _indexRebuilder.RebuildIndexes(_onlyEmptyIndexes);
             }
         }
