+++ src/Umbraco.Web/Search/ExamineFinalComponent.cs
@@ -12,15 +12,11 @@ namespace Umbraco.Web.Search
     /// </summary>
     public sealed class ExamineFinalComponent : IComponent
     {   
-        private readonly IProfilingLogger _logger;
-        private readonly IExamineManager _examineManager;
         BackgroundIndexRebuilder _indexRebuilder;
         private readonly IMainDom _mainDom;
         
-        public ExamineFinalComponent(IProfilingLogger logger, IExamineManager examineManager, BackgroundIndexRebuilder indexRebuilder, IMainDom mainDom)
+        public ExamineFinalComponent(BackgroundIndexRebuilder indexRebuilder, IMainDom mainDom)
         {
-            _logger = logger;
-            _examineManager = examineManager;
             _indexRebuilder = indexRebuilder;
             _mainDom = mainDom;
         }
@@ -28,8 +24,6 @@ namespace Umbraco.Web.Search
         public void Initialize()
         {
             if (!_mainDom.IsMainDom) return;
-
-            _examineManager.ConfigureIndexes(_mainDom, _logger);
 
             // TODO: Instead of waiting 5000 ms, we could add an event handler on to fulfilling the first request, then start?
             _indexRebuilder.RebuildIndexes(true, 5000);
