+++ src/Umbraco.Web/Umbraco.Web.csproj
@@ -65,9 +65,6 @@
     <PackageReference Include="Examine.Core">
       <Version>2.0.0-alpha.20200128.15</Version>
     </PackageReference>
-    <PackageReference Include="Examine.Lucene">
-      <Version>2.0.0-alpha.20200128.15</Version>
-    </PackageReference>
     <PackageReference Include="HtmlAgilityPack" Version="1.8.14" />
     <PackageReference Include="ImageProcessor">
       <Version>2.7.0.100</Version>
@@ -115,10 +112,6 @@
     <ProjectReference Include="..\Umbraco.Examine\Umbraco.Examine.csproj">
       <Project>{f9b7fe05-0f93-4d0d-9c10-690b33ecbbd8}</Project>
       <Name>Umbraco.Examine</Name>
-    </ProjectReference>
-    <ProjectReference Include="..\Umbraco.Examine.Lucene\Umbraco.Examine.Lucene.csproj">
-      <Project>{07fbc26b-2927-4a22-8d96-d644c667fecc}</Project>
-      <Name>Umbraco.Examine.Lucene</Name>
     </ProjectReference>
     <ProjectReference Include="..\Umbraco.Infrastructure\Umbraco.Infrastructure.csproj">
       <Project>{3ae7bf57-966b-45a5-910a-954d7c554441}</Project>
@@ -246,7 +239,6 @@
     <Compile Include="Security\UmbracoEmailMessage.cs" />
     <Compile Include="Security\UmbracoMembershipProviderBase.cs" />
     <Compile Include="Security\UserAwarePasswordHasher.cs" />
-    <Compile Include="Search\IUmbracoTreeSearcherFields.cs" />
     <Compile Include="Search\UmbracoTreeSearcherFields.cs" />
     <Compile Include="StringExtensions.cs" />
     <Compile Include="Templates\HtmlLocalLinkParser.cs" />
@@ -271,9 +263,6 @@
     <Compile Include="Runtime\WebInitialComposer.cs" />
     <Compile Include="Scheduling\SchedulerComposer.cs" />
     <Compile Include="Search\ExamineComposer.cs" />
-    <Compile Include="Search\GenericIndexDiagnostics.cs" />
-    <Compile Include="Search\IUmbracoIndexesCreator.cs" />
-    <Compile Include="Search\UmbracoIndexesCreator.cs" />
     <Compile Include="Security\ActiveDirectoryBackOfficeUserPasswordChecker.cs" />
     <Compile Include="Security\BackOfficeClaimsIdentityFactory.cs" />
     <Compile Include="Security\BackOfficeUserManagerMarker.cs" />
